Fixes #568: Add `rejectUnauthorized` to option for https.request
Adding `rejectUnauthorized` in the request enables unverified connections to
secure hosts. In older versions, it seems to work via the https.request.agent,
but it is ignored in newer versions, specifically the latest version of
VS Code's extension executions.

The primary use case is mlxprs eval through a bastion SSH tunnel into a private
DHS instance. In the current state, the request is rejected regardless of the
user's `rejectUnauthorized` setting.